// src/app/pages/admin/model/traindetails.ts
export class traindetails {
  tid: number = 0;
  tno: number = 0;
  tname: string = '';
  start: string = '';
  stop: string = '';
  date: any = '';	//take care of format
  totalCoach: number = 0;
  acSleeperCoach: number = 0;
  acSittingCoach: number = 0;
  nonAcSleeperCoach: number = 0;
  nonAcSittingCoach: number = 0;
  totalAcSleeperSeat: number = 0;
  availAcSleeperSeat: number = 0;
  totalAcSittingSeat: number = 0;
  availAcSittingSeat: number = 0;
  totalNonAcSleeperSeat: number = 0;
  availNonAcSleeperSeat: number = 0;
  totalNonAcSittingSeat: number = 0;
  availNonAcSittingSeat: number = 0;
}